% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{dpmirt_compare}
\alias{dpmirt_compare}
\title{Compare DPMirt models using information criteria}
\usage{
dpmirt_compare(..., criterion = "waic")
}
\arguments{
\item{...}{Two or more \code{dpmirt_fit} objects.}

\item{criterion}{Character. Comparison criterion. Default "waic".}
}
\value{
A data.frame ranking models by the criterion.
}
\description{
Compare DPMirt models using information criteria
}
\examples{
\dontrun{
sim <- dpmirt_simulate(200, 20, model = "rasch", seed = 42)
fit1 <- dpmirt(sim$response, model = "rasch", prior = "normal",
               niter = 5000, nburnin = 1000, seed = 123)
fit2 <- dpmirt(sim$response, model = "rasch", prior = "dpm",
               niter = 5000, nburnin = 1000, seed = 123)

# Compare via WAIC
comp <- dpmirt_compare(fit1, fit2)
print(comp)
}

}
\seealso{
\code{\link{dpmirt}}, \code{\link{dpmirt_diagnostics}}

Other diagnostics: 
\code{\link{dpmirt_diagnostics}()}
}
\concept{diagnostics}
