% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg.R
\name{dpmirt_plot_density}
\alias{dpmirt_plot_density}
\title{Plot Posterior Mean Density}
\usage{
dpmirt_plot_density(fit, ...)
}
\arguments{
\item{fit}{A \code{dpmirt_fit} object from \code{\link{dpmirt}}.}

\item{...}{Currently unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Displays a kernel density estimate of the posterior mean person abilities
(\eqn{\hat{\theta}^{PM}}). Useful for assessing the shape of the estimated
latent trait distribution. Requires ggplot2.
}
\examples{
\dontrun{
sim <- dpmirt_simulate(200, 20, model = "rasch", seed = 42)
fit <- dpmirt(sim$response, model = "rasch", prior = "normal",
              niter = 5000, nburnin = 1000, seed = 123)
dpmirt_plot_density(fit)
}

}
\seealso{
\code{\link{plot.dpmirt_fit}}, \code{\link{dpmirt_plot_density_compare}}

Other visualization: 
\code{\link{dpmirt_plot_caterpillar}()},
\code{\link{dpmirt_plot_clusters}()},
\code{\link{dpmirt_plot_density_compare}()},
\code{\link{dpmirt_plot_dp_density}()},
\code{\link{dpmirt_plot_icc}()},
\code{\link{dpmirt_plot_info}()},
\code{\link{dpmirt_plot_items}()},
\code{\link{dpmirt_plot_parameter_trace}()},
\code{\link{dpmirt_plot_pp_check}()},
\code{\link{dpmirt_plot_trace}()},
\code{\link{dpmirt_plot_wright_map}()},
\code{\link{plot.dpmirt_estimates}()},
\code{\link{plot.dpmirt_fit}()},
\code{\link{plot.dpmirt_sim}()}
}
\concept{visualization}
