% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dpmirt_sim}
\alias{plot.dpmirt_sim}
\title{Plot dpmirt_sim object}
\usage{
\method{plot}{dpmirt_sim}(x, type = c("parameters", "response"), ...)
}
\arguments{
\item{x}{A \code{dpmirt_sim} object.}

\item{type}{Character. Plot type: \code{"parameters"} (histograms of true
parameters) or \code{"response"} (heatmap of response matrix).}

\item{...}{Additional graphical parameters.}
}
\description{
Plot dpmirt_sim object
}
\examples{
\dontrun{
sim <- dpmirt_simulate(200, 20, model = "rasch", seed = 42)

# True parameter histograms
plot(sim, type = "parameters")

# Response matrix heatmap
plot(sim, type = "response")
}

}
\seealso{
\code{\link{dpmirt_simulate}}

Other visualization: 
\code{\link{dpmirt_plot_caterpillar}()},
\code{\link{dpmirt_plot_clusters}()},
\code{\link{dpmirt_plot_density}()},
\code{\link{dpmirt_plot_density_compare}()},
\code{\link{dpmirt_plot_dp_density}()},
\code{\link{dpmirt_plot_icc}()},
\code{\link{dpmirt_plot_info}()},
\code{\link{dpmirt_plot_items}()},
\code{\link{dpmirt_plot_parameter_trace}()},
\code{\link{dpmirt_plot_pp_check}()},
\code{\link{dpmirt_plot_trace}()},
\code{\link{dpmirt_plot_wright_map}()},
\code{\link{plot.dpmirt_estimates}()},
\code{\link{plot.dpmirt_fit}()}
}
\concept{visualization}
